requirements.txt

# Advanced ML
# AI & Machine Learning (Enhanced)
# AI/ML Dependencies
aiohttp>=3.12.13
aiohttp==3.9.1
# AI TRADING SYSTEM REQUIREMENTS
alpha-vantage==2.3.1
# API & Web
arch==6.2.0
# Async & HTTP
asyncio>=3.4.3
asyncio-mqtt==0.16.1
# Async Support
bcrypt==4.1.2
black==23.11.0
catboost>=1.0.0
ccxt==4.2.0
ccxt>=4.4.0,<5.0.0
ccxt>=4.4.90
click==8.1.7
colorama==0.4.6
# Configuration & Utilities
# Core AI/ML Dependencies
# Core Trading Dependencies
# Core Trading & Exchange
cryptography==41.0.8
cvxpy>=1.4.0
cython==3.0.6
# Database & Storage
# Data Processing & Analysis
# Data Science
# Data Validation & Config
# Deep Learning Utils
# Development
# Environment & System
# Exchange & Trading (kompatible Version)
fastapi==0.104.1
# Financial Data
flake8==6.1.0
gym>=0.21.0
gym>=0.26.0
gym==0.26.2
gymnasium==0.29.1
hydra-core>=1.1.0
hyperopt==0.2.7
ipython==8.17.2
joblib>=1.0.0
joblib>=1.1.0
joblib==1.3.2
jupyter==1.0.0
# Jupyter & Development
keras==2.15.0
lightgbm>=3.3.0
# Logging
# Logging & Monitoring
loguru>=0.6.0
loguru==0.7.2
loguru>=0.7.3
# Machine Learning & AI
marshmallow==3.20.2
matplotlib>=3.5.0 
matplotlib==3.8.2
nltk>=3.7
notebook==7.0.6
numba>=0.56.0
numba==0.58.1
numpy>=1.21.0,<2.0.0
numpy==1.25.2
numpy>=2.3.1
omegaconf>=2.1.0
# Optional GPU Support (falls verfügbar)
optuna>=2.10.0
optuna==3.5.0
pandas>=1.3.0,<3.0.0
pandas==2.1.4
pandas>=2.3.0
pandas-ta>=0.3.14b
# Performance
# Performance & Optimization
plotly==5.17.0
plotly>=5.8.0
pmdarima==2.0.4
# Portfolio Optimization
psutil==5.9.6
psutil>=7.0.0
pydantic>=1.9.0
pydantic>=2.11.7
pydantic==2.5.2
pytest>=6.0.0 # requirements.txt - Kompatible Dependencies für Python 3.12
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv>=0.19.0
python-dotenv==1.0.0
python-dotenv>=1.1.0
python-telegram-bot==20.7
python-telegram-bot>=22.0
python-telegram-bot>=22.1
pytorch-lightning==2.1.3
pytz==2023.3
PyYAML==6.0.1
PyYAML>=6.0.2
quandl==3.7.0
ray[tune]==2.8.1
redis==5.0.1
requests==2.31.0
# RL Dependencies - Reinforcement Learning System
sb3-contrib==2.2.1
schedule==1.2.0 
scikit-learn>=1.0.0
scikit-learn>=1.1.0
scikit-learn==1.3.2
scikit-learn>=1.7.0
scipy==1.11.4
scipy>=1.7.0
scipy>=1.9.0
seaborn>=0.11.0
seaborn==0.13.0
# Security & Encryption
# Sentiment Analysis
shimmy>=2.0.0
sqlalchemy==2.0.23
stable-baselines3>=1.6.0
stable-baselines3>=2.0.0
stable-baselines3==2.2.1
statsmodels==0.14.1
structlog==23.2.0
ta>=0.10.0
ta==0.10.2
ta>=0.10.2
ta>=0.11.0
ta-lib==0.4.28
talib-binary>=0.4.19
# Technical Analysis
# Technical Analysis Enhanced
telegram==0.0.1
# Telegram Bot
tensorboard>=2.12.0
tensorboard==2.15.1
tensorflow==2.15.0
# tensorflow>=2.8.0
tensorflow>=2.8.0
# Testing & Development
textblob>=0.17.1
# Time Series & Financial
torch>=1.11.0
torch>=1.11.0,<2.5.0
torch>=2.0.0
torch==2.1.2
torch-audio==2.1.2
torchvision>=0.12.0
torchvision>=0.12.0,<0.20.0
torch-vision==0.16.2
tqdm>=4.60.0
tqdm==4.66.1
# TRADINO UNSCHLAGBAR - AI DEPENDENCIES
# Utilities
uvicorn==0.24.0
vaderSentiment>=3.3.2
# Visualization Enhanced
# Visualization & Plotting
wandb==0.16.1
websockets==12.0
xgboost>=1.6.0
xgboost==2.0.3
xgboost>=3.0.2
yfinance==0.2.18
